src/lesson43/setImplemenration/OurHashSet.java
package lesson43.setImplemenration;

import java.util.HashMap;
import java.util.Iterator;
import java.util.Map;

public class OurHashSet<E> implements OurSet<E> {

    private Map<E, Object> source;
    private final static Object INSTANCE = new Object();

    public OurHashSet() {
        this.source = new HashMap<>();
    }

    @Override
    public boolean add(E elt) {
        return source.put(elt, INSTANCE) == null;
    }

    @Override
    public boolean remove(E elt) {
        return source.remove(elt) != null;
    }

    @Override
    public boolean contains(E elt) {
        return source.containsKey(elt);
    }

    @Override
    public int size() {
        return source.size();
    }

    @Override
    public Iterator<E> iterator() {
        return source.keySet().iterator();
    }
}
